Minor refactor and added some doc strings
